feat(walkthrough): add shape rotate animation based on scroll view gesture
